backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.40.0
faiss-cpu>=1.8.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.7.0

# File parsing
pypdf>=4.0.0
python-docx>=1.1.0

# LangChain conversation memory
langchain-classic>=1.0.0
langchain-core>=0.3.0
python-multipart>=0.0.12
